+++ src/datasources/exchange-api/exchange.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { HttpErrorHandler } from '../errors/http-error-handler';
-import { ExchangeService } from './exchange.service';
+import { ExchangeApi } from './exchange.service';
 
 @Module({
-  providers: [HttpErrorHandler, ExchangeService],
-  exports: [ExchangeService],
+  providers: [HttpErrorHandler, ExchangeApi],
+  exports: [ExchangeApi],
 })
 export class ExchangeModule {}
